Split up ctffind
